Update platform image
This image suddenly caused transparency issues with the game engine, so I changed the borders from black to dark green.

The game automatically assigns black as an image's transparent color if it's the first color in the image's palette. I don't know why the black color didn't appear transparent before, but it does now.